Fix protocol validation, USB alt setting, PIT header, ZLP read, compressed download support, and split usb_device.cpp into modules
- odin_fail_check: validate full 4-byte id field instead of just first byte,
  check for BOOTLOADER_FAIL (0xFFFFFFFF), verify ack < 0 as error indicator,
  accept optional expected_id parameter
- Add libusb_set_interface_alt_setting after claiming interface to ensure
  correct alternate setting is used
- Fix PIT header fields: identify com_tar2, cpu_bl_id, lu_count instead
  of unknown1-8, enabling device model validation
- Fix ZLP consumption in odin_dump_pit: use zero-length bulk transfer
  instead of reading 64 bytes
- Add compressed download (RQT_XMIT_COMPRESSED) support when device
  advertises capability via ack upper bit 0x8000 in BeginSession response
- Split usb_device.cpp into usb_device.cpp (USB transport/enumeration)
  and odin_protocol.cpp (protocol commands, PIT parsing, flash methods)
