+++ src/CMakeLists.txt
@@ -3,5 +3,6 @@ set(ODIN4_CORE_SOURCES
     ${CMAKE_CURRENT_SOURCE_DIR}/odin4.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/core/logger.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/usb/usb_device.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/usb/odin_protocol.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/firmware/firmware_package.cpp
     PARENT_SCOPE)
